Extract defra-node P2P runtime into private module
Move lifecycle, setup, restore, failure-recorder, and Iroh retry
responsibilities out of the 2k-line lib.rs into p2p_runtime.rs without
changing public API or task shutdown/retry semantics.
